+++ src/components/input/input.styles.ts
@@ -20,46 +20,65 @@ export default css`
     font-family: var(--sl-input-font-family);
     font-weight: var(--sl-input-font-weight);
     letter-spacing: var(--sl-input-letter-spacing);
-    background-color: rgb(var(--sl-input-background-color));
-    border: solid var(--sl-input-border-width) rgb(var(--sl-input-border-color));
     vertical-align: middle;
     overflow: hidden;
-    transition: var(--sl-transition-fast) color, var(--sl-transition-fast) border, var(--sl-transition-fast) box-shadow;
     cursor: text;
+    transition: var(--sl-transition-fast) color, var(--sl-transition-fast) border, var(--sl-transition-fast) box-shadow,
+      var(--sl-transition-fast) background-color;
   }
 
-  .input:hover:not(.input--disabled) {
+  /* Standard inputs */
+  .input--standard {
+    background-color: rgb(var(--sl-input-background-color));
+    border: solid var(--sl-input-border-width) rgb(var(--sl-input-border-color));
+  }
+
+  .input--standard:hover:not(.input--disabled) {
     background-color: rgb(var(--sl-input-background-color-hover));
     border-color: rgb(var(--sl-input-border-color-hover));
   }
 
-  .input:hover:not(.input--disabled) .input__control {
-    color: rgb(var(--sl-input-color-hover));
-  }
-
-  .input.input--focused:not(.input--disabled) {
+  .input--standard.input--focused:not(.input--disabled) {
     background-color: rgb(var(--sl-input-background-color-focus));
     border-color: rgb(var(--sl-input-border-color-focus));
     box-shadow: var(--sl-focus-ring);
   }
 
-  .input.input--focused:not(.input--disabled) .input__control {
+  .input--standard.input--focused:not(.input--disabled) .input__control {
     color: rgb(var(--sl-input-color-focus));
   }
 
-  .input.input--disabled {
+  .input--standard.input--disabled {
     background-color: rgb(var(--sl-input-background-color-disabled));
     border-color: rgb(var(--sl-input-border-color-disabled));
     opacity: 0.5;
     cursor: not-allowed;
   }
 
-  .input.input--disabled .input__control {
+  .input--standard.input--disabled .input__control {
     color: rgb(var(--sl-input-color-disabled));
   }
 
-  .input.input--disabled .input__control::placeholder {
+  .input--standard.input--disabled .input__control::placeholder {
     color: rgb(var(--sl-input-placeholder-color-disabled));
+  }
+
+  /* Filled inputs */
+  .input--filled {
+    background-color: rgb(var(--sl-color-neutral-100));
+    border: solid 1px rgb(var(--sl-color-neutral-100));
+    color: rgb(var(--sl-input-color));
+  }
+
+  .input--filled.input--focused:not(.input--disabled) {
+    background-color: rgb(var(--sl-color-neutral-0));
+    border-color: rgb(var(--sl-color-primary-500));
+    box-shadow: var(--sl-focus-ring);
+  }
+
+  .input--filled.input--disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
   }
 
   .input__control {
@@ -97,6 +116,10 @@ export default css`
   .input__control::placeholder {
     color: rgb(var(--sl-input-placeholder-color));
     user-select: none;
+  }
+
+  .input:hover:not(.input--disabled) .input__control {
+    color: rgb(var(--sl-input-color-hover));
   }
 
   .input__control:focus {